[Fix][Flink-starter]Hotfix flink checkpoint default value (#7040)
